add transition state
