+++ firefox-ios/Client/Frontend/Library/Bookmarks/BookmarksViewController.swift
@@ -146,6 +146,17 @@ class BookmarksViewController: SiteTableViewController,
         }
     }
 
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+
+        if viewModel.isRootNode {
+            updatePanelState(newState: .bookmarks(state: .mainView))
+        } else {
+            updatePanelState(newState: .bookmarks(state: .inFolder))
+        }
+        sendPanelChangeNotification()
+    }
+
     // MARK: - Data
 
     override func reloadData() {
@@ -795,16 +806,8 @@ extension BookmarksViewController {
     }
 
     func handleLeftTopButton() {
-        guard case .bookmarks(let subState) = state else { return }
-
-        switch subState {
-        case .inFolder:
-            if viewModel.isRootNode {
-                updatePanelState(newState: .bookmarks(state: .mainView))
-            }
-        default:
-            return
-        }
+        // We use the "transitioning" so that the "<" back toolbar navigation button cannot be spammed
+        updatePanelState(newState: .bookmarks(state: .transitioning))
     }
 
     func shouldDismissOnDone() -> Bool {
