+++ firefox-ios/Client/Frontend/Library/LibraryViewController/LibraryPanelViewState.swift
@@ -48,6 +48,7 @@ enum LibraryPanelSubState {
     case itemEditMode
     case itemEditModeInvalidField
     case search
+    case transitioning
 
     // The following two functions enable checking that substate moves are legal.
     // For example, you can move from .mainView to .inFolder, but not from
